renamed files and added parsing command line
